moves admin.tsx into admin/index.tsx
